+++ src/main/java/org/recordy/server/auth/exception/UnauthorizedSocialTokenException.java
@@ -3,12 +3,12 @@ package org.recordy.server.auth.exception;
 import org.recordy.server.auth.message.ErrorMessage;
 import org.recordy.server.common.exception.RecordyException;
 
-public class UnauthorizedException extends RecordyException {
-    public UnauthorizedException() {
+public class UnauthorizedSocialTokenException extends RecordyException {
+    public UnauthorizedSocialTokenException() {
         super(ErrorMessage.UNAUTHORIZED);
     }
 
-    public UnauthorizedException(ErrorMessage errorMessage) {
+    public UnauthorizedSocialTokenException(ErrorMessage errorMessage) {
         super(errorMessage);
     }
 }
